pyproject.toml
[project]
name = "trade-crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ib-async>=2.1.0",
    "loguru>=0.7.3",
    "pydantic-ai>=1.70.0",
    "python-dotenv>=1.2.2",
]
